Happy New Year 🍾🎇 Goodbye 2021. Welcome 2022

+++ Server/src/Chessbook.Services/Chessbook.Services/Blocklist/BlocklistService.cs
@@ -109,6 +109,11 @@ namespace Chessbook.Services.Blocklist
 
             var blockListModel = await this.LoadByUserAndTarget(userId, targetUser.Id);
 
+            if (blockListModel == null)
+            {
+                return;
+            }
+
             await this.userBlocklistRepository.DeleteAsync(blockListModel);
 
             // update relationship
